refactored header file for qicosahedron.cpp
